Refactor rerank_recursive.py to package re-export
Replaces the implementation in scripts/rerank_recursive.py with a backwards-compatibility shim that re-exports all symbols from the scripts.rerank_recursive package. All logic and classes have been moved to the scripts/rerank_recursive package modules, and the old script now only exists to maintain compatibility for existing imports.
